% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_class_2_colours}
\alias{get_class_2_colours}
\title{Map marker class names to colours}
\usage{
get_class_2_colours()
}
\value{
\verb{named list} with colours for each possible expected marker class
}
\description{
To ensure standardised marker colours, we define a
set of colours for the 'normal' marker sets.
}
