% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_colours}
\alias{get_colours}
\title{Get colours for a set of marker classes}
\usage{
get_colours(marker_classes, class_2_colour = get_class_2_colours())
}
\arguments{
\item{marker_classes}{\code{vector} marker class names}

\item{class_2_colour}{\verb{named list} mapping marker class names to hex colour}
}
\value{
\code{vector} of hex colour codes
}
\description{
To ensure standardised marker colours, we define a
set of colours for the 'normal' marker sets.
}
