% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_profiles.R
\name{get_sample_2_rep}
\alias{get_sample_2_rep}
\title{Return a map of sample to replicate based on MSnSet column names and
replicate boundaries}
\usage{
get_sample_2_rep(obj, replicate_boundaries)
}
\arguments{
\item{obj}{\code{MSnSet}}

\item{replicate_boundaries}{\verb{numeric vector} defining the boundaries between
replicates. E.g if 3 replicates with 10 samples each, boundaries are \code{c(10,20)}}
}
\description{
Return a map of sample to replicate based on MSnSet column names and
replicate boundaries
}
