% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proj.R
\name{make_proj}
\alias{make_proj}
\title{Make 2D projections}
\usage{
make_proj(obj, method = "PCA", fcol = "markers", dims = c(1, 2), ...)
}
\arguments{
\item{obj}{\code{MSnSet} containing features to plot}

\item{method}{\code{character} describing \code{pRoloc::plot2D} method to use}

\item{fcol}{\code{character} Feature meta-data label (fData column name) defining
the groups to be differentiated using different colours. Default is markers}

\item{dims}{A \code{numeric} of length 2 defining the dimensions to be plotted.
Default is c(1, 2)}

\item{...}{additional arguments to be passed to \code{pRoloc::plot2D}}
}
\description{
Generates 2D projections using either PCA or t-SNE
}
\examples{
library(pRolocdata)
data(lopitdcU2OS2018)
make_proj(lopitdcU2OS2018)
}
