% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_profiles.R
\name{Plot_feature_profiles}
\alias{Plot_feature_profiles}
\title{Plots linear profiles for each feature}
\usage{
Plot_feature_profiles(obj, replicate_boundaries = NULL)
}
\arguments{
\item{obj}{\code{MSnSet} containing features to plot}

\item{replicate_boundaries}{\verb{numeric vector} defining the boundaries between
replicates. E.g if 3 replicates with 10 samples each, boundaries are \code{c(10,20)}}
}
\description{
Generates line plots where each feature is plotted separately.
If \code{replicate_boundaries} are supplied, the plot is facetted by replicate.
Assumes order of samples in input MSnSet is the correct order for plotting
}
\examples{
library(pRolocdata)
data(lopitdcU2OS2018)
Plot_feature_profiles(head(lopitdcU2OS2018), c(10,20))
}
