% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_profiles.R
\name{plot_marker_profiles}
\alias{plot_marker_profiles}
\title{Plots linear profiles for each marker set}
\usage{
plot_marker_profiles(
  obj,
  fcol = "markers",
  plot_all = T,
  alpha = 1,
  facet_by = "none",
  group_by = NULL
)
}
\arguments{
\item{obj}{\code{MSnSet} containing features to plot}

\item{fcol}{\code{character} defining the markers column}

\item{plot_all}{\code{logical} should all individual features be plotted}

\item{alpha}{\code{numeric} alpha value(s) for marker sets}

\item{facet_by}{\code{character} Split plot into facets for each marker sets
(\code{markers}), each feature (\code{feature}), or no facetting (\code{none})}

\item{group_by}{Grouping factor for lines. Variable must be specified in pData for obj.
This parameter can be used to e.g ensure replicates are plotted as separate lines}
}
\description{
Generates line plots for each marker. Each marker set is plotted
in a separate facet. Assumes order of samples in input MSnSet is the correct
order for plotting
}
\examples{
library(pRolocdata)
data(lopitdcU2OS2018)
obj <- lopitdcU2OS2018[,1:10]
obj <- obj[fData(obj)$markers != 'unknown',]
alpha <- 10*classWeights(obj)
plot_marker_profiles(obj, facet_by='markers', alpha=alpha)
}
